Import Lodash and Apply Styling to Component
Description: Integrated the Lodash library by importing it in the JavaScript entry file, `index.js`. Additionally, imported the `style.css` file to apply styling to the component. The `component` function now creates a `div` element with content generated using Lodash's `_.join` function, and this element is appended to the `body` of the HTML document. This results in styled content being displayed on the webpage.