Generate Distribution Files
Built and generated distribution files for the project. Added `dist/index.html` and `dist/index.bundle.js` to version control. These files contain the bundled and optimized JavaScript code along with the HTML template, ready to be deployed to a web server or used for production. This commit marks the completion of the build process and prepares the project for deployment or further development.